requirements.txt
kumoai>=0.1.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.3.0
streamlit>=1.38.0
pandas>=2.0.0
pyarrow>=14.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
sentence-transformers>=3.0.0
numpy>=1.26.0
